chore: implement useEvalConfigFilterOptions hook for centralized filter management in evaluator form
